requirements.txt
python==3.10.11
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
